Add tabbed layout for exports
