--- src/views/UserExports/styles.scss
@@ -1,127 +1,0 @@
-.user-exports {
-    display: flex;
-    flex-grow: 1;
-
-    .header {
-        display: flex;
-        align-items: center;
-
-        .heading {
-            flex-grow: 1;
-        }
-    }
-
-    .main-content {
-        display: flex;
-        position: relative;
-
-        .table-container {
-            display: flex;
-            flex-direction: column;
-            flex-grow: 1;
-            background-color: var(--color-foreground);
-            overflow-y: auto;
-
-            .loading-animation {
-                position: relative;
-                height: var(--font-size-large);
-
-                >div {
-                    background-color: transparent;
-                    padding: 0!important;
-                    font-size: var(--font-size-large)!important;
-                }
-            }
-
-            .table {
-                flex-grow: 1;
-
-                :global {
-                    .file-download {
-                        display: flex;
-                        justify-content: center;
-                        color: var(--color-accent);
-                    }
-
-                    .table {
-                        .headers {
-                            .mime-type {
-                                width: 96px;
-                            }
-
-                            .file {
-                                width: 64px;
-                            }
-                        }
-
-                        .body {
-                            cursor: pointer;
-
-                            .row {
-                                &.highlighted {
-                                    color: var(--color-text-on-accent);
-
-                                    .file-download {
-                                        color: var(--color-text-on-accent);
-                                    }
-                                }
-
-                                &:hover {
-                                    background-color: var(--color-background-accent-hint);
-                                    color: var(--color-text);
-
-                                    .file-download {
-                                        color: var(--color-accent);
-                                    }
-                                }
-                            }
-
-                            .cell {
-                                position: relative;
-                            }
-
-                            .button {
-                                &:not(:last-child) {
-                                    margin-right: var(--spacing-medium);
-                                }
-                            }
-                        }
-                    }
-
-                    .icon-wrapper {
-                        display: flex;
-                        justify-content: center;
-                    }
-
-                    .to-be-started {
-                        display: flex;
-                        justify-content: center;
-                    }
-
-                    .file-error {
-                        display: flex;
-                        justify-content: center;
-                        color: var(--color-danger);
-                    }
-                }
-            }
-        }
-
-        .preview {
-            flex-basis: 30%;
-            flex-grow: 0;
-            flex-shrink: 0;
-            border-left: var(--width-separator-thin) solid var(--color-separator);
-            background-color: var(--color-foreground-alt);
-        }
-    }
-
-    .footer {
-        display: flex;
-        justify-content: flex-end;
-
-        .pager {
-            padding: var(--spacing-extra-small) 0;
-        }
-    }
-}
